requirements.txt
# API FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart>=0.0.5

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
scipy>=1.11.0

# Cloud & Monitoring
google-cloud-monitoring==2.16.0
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0
kfp>=2.0.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
psutil>=5.9.0

# MLflow pour gestion des modeles et experimentations
mlflow>=2.8.0
mlflow[extras]>=2.8.0

# Utilitaires
python-dotenv==1.0.0
requests==2.31.0

# Sécurité & JWT
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Tests
pytest>=7.0.0
pytest-asyncio>=0.21.0